PythonTest/requirements.txt
base58==2.1.0
ecdsa==0.16.1
python-bitcoinlib==0.11.0
six==1.15.0
#pip3 install -r requirements.txt